+++ single_channel_player_app/lib/widgets/pulse_loader.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_animate/flutter_animate.dart';
+import 'dart:math' as math;
 
-class PulseLoader extends StatelessWidget {
+class PulseLoader extends StatefulWidget {
   final Color color;
   final double size;
 
@@ -12,33 +12,83 @@ class PulseLoader extends StatelessWidget {
   });
 
   @override
+  State<PulseLoader> createState() => _PulseLoaderState();
+}
+
+class _PulseLoaderState extends State<PulseLoader>
+    with SingleTickerProviderStateMixin {
+  late AnimationController _controller;
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = AnimationController(
+      vsync: this,
+      duration: const Duration(milliseconds: 900),
+    )..repeat();
+  }
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    final double dotSize = size * 0.22;
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      children: List.generate(3, (i) {
-        return Container(
-          margin: EdgeInsets.symmetric(horizontal: dotSize * 0.35),
-          width: dotSize,
-          height: dotSize,
-          decoration: BoxDecoration(color: color, shape: BoxShape.circle),
-        )
-            .animate(onPlay: (c) => c.repeat())
-            .moveY(
-              begin: 0,
-              end: -(size * 0.35),
-              delay: Duration(milliseconds: i * 160),
-              duration: 380.ms,
-              curve: Curves.easeInOut,
-            )
-            .then()
-            .moveY(
-              begin: -(size * 0.35),
-              end: 0,
-              duration: 380.ms,
-              curve: Curves.easeInOut,
-            );
-      }),
+    return AnimatedBuilder(
+      animation: _controller,
+      builder: (_, __) => Transform.rotate(
+        angle: _controller.value * 2 * math.pi,
+        child: CustomPaint(
+          size: Size(widget.size, widget.size),
+          painter: _RingPainter(
+            color: widget.color,
+            strokeWidth: widget.size * 0.1,
+          ),
+        ),
+      ),
     );
   }
+}
+
+class _RingPainter extends CustomPainter {
+  final Color color;
+  final double strokeWidth;
+
+  const _RingPainter({required this.color, required this.strokeWidth});
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    final center = Offset(size.width / 2, size.height / 2);
+    final radius = (size.width - strokeWidth) / 2;
+
+    // Faint background track
+    canvas.drawCircle(
+      center,
+      radius,
+      Paint()
+        ..color = color.withValues(alpha: 0.15)
+        ..strokeWidth = strokeWidth
+        ..style = PaintingStyle.stroke
+        ..strokeCap = StrokeCap.round,
+    );
+
+    // 270° spinning arc
+    canvas.drawArc(
+      Rect.fromCircle(center: center, radius: radius),
+      -math.pi / 2,
+      3 * math.pi / 2,
+      false,
+      Paint()
+        ..color = color
+        ..strokeWidth = strokeWidth
+        ..style = PaintingStyle.stroke
+        ..strokeCap = StrokeCap.round,
+    );
+  }
+
+  @override
+  bool shouldRepaint(_RingPainter old) =>
+      old.color != color || old.strokeWidth != strokeWidth;
 }